requirements.txt
Flask==3.0.3
pymongo==4.7.2
requests==2.32.3
python-dotenv==1.0.1
